services/alert-ingestion/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pika==1.3.2
python-dotenv==1.0.0
prometheus-client==0.19.0
